refactor(client): keep row draft state local to the capabilities section
